Refactor Playlist.cs into FileData directory.
