+++ MusicSharpTests/PlaylistLoaderTests.cs
@@ -1,4 +1,4 @@
-using MusicSharp;
+using MusicSharp.FileData;
 
 namespace MusicSharpTests;
